api/requirements.txt
pytube==12.1.0
pydub==0.25.1
python-dotenv==0.20.0
ibm-watson==6.0.0
Flask==2.1.2
Werkzeug==2.1.2
gunicorn==20.1.0
mongoengine==0.24.1
dnspython==2.2.1